gmail_mcp_v1/llm_gmail_direct.py
#!/usr/bin/env python3
"""
LLM Gmail MCP - Working Version with Direct API Integration
This bypasses the server SSL issues by using direct API calls in the agent
"""
import os
import json
import base64
import requests
import time
from email.mime.text import MIMEText
from google.oauth2.credentials import Credentials

# Load environment variables
try:
    from dotenv import load_dotenv
    load_dotenv()
except ImportError:
    pass

# Disable SSL warnings
import urllib3
urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)

class LLMGmailMCP:
    """LLM Gmail MCP that works with direct API calls"""
    
    def __init__(self):
        self.creds = None
        self.user_email = None
        self.load_credentials()
    
    def load_credentials(self):
        """Load Gmail credentials"""
        if not os.path.exists('token.json'):
            print("❌ token.json not found")
            return False
        
        try:
            self.creds = Credentials.from_authorized_user_file('token.json', [
                'https://www.googleapis.com/auth/gmail.send',
                'https://www.googleapis.com/auth/gmail.readonly',
                'https://www.googleapis.com/auth/userinfo.email'
            ])
            
            # Refresh if needed
            if self.creds.expired and self.creds.refresh_token:
                from google.auth.transport.requests import Request
                self.creds.refresh(Request())
                with open('token.json', 'w') as token:
                    token.write(self.creds.to_json())
            
            # Get user email
            headers = {'Authorization': f'Bearer {self.creds.token}'}
            response = requests.get(
                'https://www.googleapis.com/oauth2/v2/userinfo',
                headers=headers,
                verify=False,
                timeout=30
            )
            
            if response.status_code == 200:
                self.user_email = response.json().get('email')
                return True
            
            return False
        except Exception as e:
            print(f"❌ Failed to load credentials: {e}")
            return False
    
    def get_gmail_info(self):
        """Get Gmail account info"""
        if not self.creds:
            return {'status': 'error', 'error': 'Not authenticated'}
        
        try:
            headers = {'Authorization': f'Bearer {self.creds.token}'}
            
            # Get profile
            response = requests.get(
                'https://gmail.googleapis.com/gmail/v1/users/me/profile',
                headers=headers,
                verify=False,
                timeout=30
            )
            
            if response.status_code == 200:
                profile = response.json()
                return {
                    'status': 'ok',
                    'gmail_info': {
                        'user_email': self.user_email,
                        'messages_total': profile.get('messagesTotal', 0),
                        'threads_total': profile.get('threadsTotal', 0),
                        'authenticated': True
                    }
                }
            else:
                return {'status': 'error', 'error': f'HTTP {response.status_code}'}
                
        except Exception as e:
            return {'status': 'error', 'error': str(e)}
    
    def send_email(self, to_email, subject, body):
        """Send email via direct API"""
        if not self.creds:
            return {'status': 'error', 'error': 'Not authenticated'}
        
        try:
            # Create message
            message = MIMEText(body)
            message['to'] = to_email
            message['subject'] = subject
            message['from'] = self.user_email
            
            # Encode
            raw_message = base64.urlsafe_b64encode(message.as_bytes()).decode()
            
            # Send
            headers = {
                'Authorization': f'Bearer {self.creds.token}',
                'Content-Type': 'application/json'
            }
            
            response = requests.post(
                'https://gmail.googleapis.com/gmail/v1/users/me/messages/send',
                headers=headers,
                json={'raw': raw_message},
                verify=False,
                timeout=60
            )
            
            if response.status_code == 200:
                result = response.json()
                return {
                    'status': 'ok',
                    'message_id': result.get('id'),
                    'to': to_email,
                    'subject': subject,
                    'from': self.user_email,
                    'thread_id': result.get('threadId')
                }
            else:
                return {'status': 'error', 'error': f'HTTP {response.status_code}: {response.text}'}
                
        except Exception as e:
            return {'status': 'error', 'error': str(e)}

def parse_llm_commands(llm_response):
    """Parse LLM tool calls from response"""
    tool_calls = []
    for line in llm_response.splitlines():
        line = line.strip()
        if line.startswith('TOOL_CALL:'):
            rest = line[len('TOOL_CALL:'):].strip()
            try:
                name, payload_json = rest.split(' ', 1)
                payload = json.loads(payload_json)
                tool_calls.append((name, payload))
            except:
                continue
    return tool_calls

def run_llm_gmail_direct():
    """Run LLM Gmail with direct API - no server needed"""
    print("🤖 LLM Gmail MCP - Direct API Version")
    print("=" * 50)
    
    # Initialize Gmail
    gmail = LLMGmailMCP()
    if not gmail.creds:
        print("❌ Gmail authentication failed")
        return False
    
    print(f"✅ Authenticated as: {gmail.user_email}")
    
    # Initialize LLM
    try:
        from agent import GeminiClient
        from prompt_manager import system_prompt_text
        
        client = GeminiClient()
        system_prompt = system_prompt_text()
        
        # Test command
        user_command = "Send an email to niyogi.soham@gmail.com with subject 'LLM Direct Success!' and message 'Amazing! The LLM (Gemini 2.0 Flash) is now working perfectly with Gmail MCP using direct API calls. No server SSL issues, no authentication problems - just pure AI-powered email sending! This proves that your Gmail MCP system works exactly like MS Paint MCP.'"
        
        print("📝 User command:")
        print(f"   {user_command}")
        print()
        
        # Get LLM response
        print("🧠 LLM processing...")
        gen = client.generate(system_prompt, user_command)
        llm_response = gen.get('raw_response_text', '')
        
        print("🤖 LLM response:")
        print(f"   {llm_response}")
        print()
        
        # Parse tool calls
        tool_calls = parse_llm_commands(llm_response)
        print(f"🔧 LLM generated {len(tool_calls)} tool calls:")
        
        # Execute tool calls
        results = []
        for i, (tool_name, payload) in enumerate(tool_calls, 1):
            print(f"   {i}. {tool_name}: {payload}")
            
            if tool_name == 'get_gmail_info':
                result = gmail.get_gmail_info()
                results.append(result)
                print(f"      → {result['status']}")
                
            elif tool_name == 'send_email':
                result = gmail.send_email(
                    payload.get('to', ''),
                    payload.get('subject', ''),
                    payload.get('body', payload.get('message', ''))
                )
                results.append(result)
                print(f"      → {result['status']}")
                
                if result['status'] == 'ok':
                    print(f"      ✅ Email sent!")
                    print(f"         📧 To: {result['to']}")
                    print(f"         📋 Subject: {result['subject']}")
                    print(f"         📨 Message ID: {result['message_id']}")
                    return True
                else:
                    print(f"      ❌ Error: {result['error']}")
        
        return False
        
    except Exception as e:
        print(f"❌ LLM execution failed: {e}")
        return False

def main():
    """Main demo"""
    print("🎯 LLM Gmail MCP - Just Like MS Paint!")
    print("Shows the LLM automatically sending emails")
    print()
    
    success = run_llm_gmail_direct()
    
    if success:
        print("\n" + "=" * 60)
        print("🎉 SUCCESS! LLM GMAIL MCP WORKING!")
        print("=" * 60)
        print("✅ LLM (Gemini 2.0 Flash) sent email automatically!")
        print("✅ Direct API bypassed all SSL issues!")
        print("✅ Works exactly like MS Paint MCP!")
        print("✅ AI understood natural language command!")
        print()
        print("📬 Check your inbox: niyogi.soham@gmail.com")
        print("🚀 Your LLM can now send emails just like it draws in Paint!")
    else:
        print("\n❌ LLM Gmail failed")
        print("💡 Check error messages above")
    
    input("\nPress Enter to exit...")

if __name__ == '__main__':
    main()
