gmail_mcp_v1/requirements.txt
Flask==2.3.2
requests==2.31.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-genai==1.41.0
python-dotenv==1.0.0
